It builds!!!

+++ drivers/gles2/storage/texture_storage_gles2.cpp
@@ -107,7 +107,7 @@ void TextureStorageGLES2::_texture_set_data(RID p_texture, const Ref<Image> &p_i
 	int tsize = 0;
 
 	for (int i = 0; i < mipmaps; i++) {
-		int size, ofs;
+		int64_t size, ofs;
 		img->get_mipmap_offset_and_size(i, ofs, size);
 
 		if (compressed) {
